refactor(app): decompose god object into focused sub-modules
Split src/app.rs (2,400 lines) into src/app/ module directory with
14 impl sub-files (navigation, layout, mouse, sort, search, rename,
file_ops, content, bookmarks, find, filter, metadata, yank, preview).

Split src/main.rs into src/args.rs (arg parsing + 9 tests),
src/shell.rs (shell integration), src/events.rs (TUI event loop).

Unit tests moved to src/app/tests.rs. src/app/mod.rs reduced to
~607 lines containing only type definitions and core lifecycle methods.

No behaviour changes; all 96 tests pass.

Closes #13
